+++ RELEASE_NOTES.md
@@ -61,6 +61,9 @@ irrelevant output.
 Installers and installation scripts
 -----------------------------------
 
+* UPDATED: The sample embedding scripts to use the `paths_core.pl` minimal file
+instead of the full `paths.pl` file.
+
 * FIXED: The `logtalk_user_setup.sh` and the Windows installer to also create
 links to the `ACKNOWLEDGMENTS.md` and `CONTRIBUTING.md` files.